airbyte-integrations/connectors/source-azure-devops/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-azure-devops"
description = "Source implementation for azure-devops."
authors = [ "Airbyte <contact@airbyte.io>", "Louis AUNEAU <louis@inarix.com>"]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/azure-devops"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_azure_devops" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
source-azure-devops = "source_azure_devops.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
freezegun = "^1.5.1"

